Feature: adicionando camada de autenticação, por enquanto esta incompleta

+++ src/main/java/Rafael/projeto_crud_clientes/controller/UsersController.java
@@ -1,9 +1,8 @@
 package Rafael.projeto_crud_clientes.controller;
 
-import Rafael.projeto_crud_clientes.entity.Users;
+import Rafael.projeto_crud_clientes.entity.User.Users;
 import Rafael.projeto_crud_clientes.service.UsersService;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,18 +20,18 @@ public class UsersController {
     }
 
     @GetMapping("/find-User-id")
-    public ResponseEntity<Users> findUserId(@RequestParam long indent) {
+    public ResponseEntity<Users> findUserId(@RequestParam Integer indent) {
         return ResponseEntity.ok(usersService.findById(indent));
     }
 
     @PutMapping("/update-user")
-    public ResponseEntity<Void> upadateUser(@RequestParam long id, @RequestBody Users user) {
+    public ResponseEntity<Void> upadateUser(@RequestParam Integer id, @RequestBody Users user) {
         usersService.updateUser(id, user);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/delete-user")
-    public ResponseEntity<Void> deleteById(@RequestParam long id) {
+    public ResponseEntity<Void> deleteById(@RequestParam Integer id) {
         usersService.deleteUserById(id);
         return ResponseEntity.ok().build();
     }
